python/agents/brand-aligned-presentations/pyproject.toml
# Copyright 2026 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.


[project]
name = "brand-aligned-presentations"
authors = [
    {name = "Renee Zhang", email = "xiaolizhang@google.com"},
    {name = "Nishant Nayak", email = "nishantnayak@google.com"}
]
version = "0.1.0"
description = "Generate a brand-adherent deck that integrates custom research."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11,<3.14"

dependencies = [
    "google-adk>=1.27.4",
    "google-auth>=2.45.0",
    "google-cloud-aiplatform[agent-engines]>=1.142.0",
    "google-cloud-modelarmor>=0.4.0",
    "python-pptx==1.0.2",
    "pillow==11.3.0",
    "pydantic>=2.12.0",
    "pydantic-settings==2.13.1",
    "python-dotenv==1.1.1",
    "google-cloud-storage==2.19.0",
    "google-genai>=1.68.0",
    "matplotlib==3.10.6",
    "opentelemetry-instrumentation-google-genai",
]

[project.optional-dependencies]
deployment = [
    "absl-py>=2.3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "uvicorn>=0.37.0",
    "httpx>=0.28.1",
    "requests>=2.32.5",
    "agent-starter-pack>=0.14.1",
    "google-adk[eval]>=1.27.4",
]

# This configuration file is used by goo.gle/agent-starter-pack to power remote templating.
# It defines the template's properties and settings.
[tool.agent-starter-pack]
agent_module = "presentation_agent.agent"
agent_object = "root_agent"

[tool.agent-starter-pack.settings]
agent_directory = "presentation_agent"
deployment_targets = ["agent_engine"]
extra_directories = ["eval", "docs"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["eval", "presentation_agent"]

[tool.ruff]
extend = "../../../pyproject.toml"
target-version = "py311"

[tool.ruff.lint]
extend-ignore = [
    "C901",    # Complexity
    "PLR0911", # Too many return statements
    "PLR0912", # Too many branches
    "PLR0915", # Too many statements
    "PLR2004", # Magic values
    "PLW0603", # Global statement
    "PLC0415", # Import not at top-level
    "RUF005",  # Ignore iterable unpacking suggestions
    "B905",    # Ignore zip() without strict parameter
    "E402",    # Ignore module level imports not at top of file (config order)
]

[tool.ruff.lint.isort]
known-first-party = ["eval", "presentation_agent"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[tool.uv]
# Universal platform support

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
python_files = ["test_*.py"]
# Suggested fix for import file mismatch
addopts = "--import-mode=importlib --cache-clear"
pythonpath = ["."]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true